model_training/requirements.txt
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
boto3==1.38.23
botocore==1.38.23
chronos==0.3
click==8.2.1
cmdstanpy==1.2.5
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
fastapi==0.115.12
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.5.1
h11==0.16.0
holidays==0.73
idna==3.10
importlib-resources==6.5.2
iniconfig==2.1.0
isort==6.0.1
jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
mako==1.3.10
markupsafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
optuna==4.3.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pip==25.1.1
platformdirs==4.3.8
pluggy==1.6.0
prophet==1.1.6
pyarrow==20.0.0
pydantic==2.11.5
pydantic-core==2.33.2
pylint==3.3.7
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
s3fs==0.4.2
s3transfer==0.13.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
stanio==0.5.1
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tomlkit==0.13.2
torch==2.7.0
tqdm==4.67.1
typing-extensions==4.13.2
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2